==> setup/03_splitting.R <==
# Final Project ----
# Initial data splitting, & data folding

# load packages ----
library(tidyverse)
library(tidymodels)
library(here)

# handle common conflicts
tidymodels_prefer()

#load data
realestate <- read_delim("data/Real_Estate_Dataset.csv", 
                         delim = ";", escape_double = FALSE, trim_ws = TRUE) |> 
  clean_names()

load(here("data/clean_qol.rda"))

#set.seed
set.seed(1234)

#split data
estate_split <- initial_split(clean_qol, 
                              prop = 0.8, 
                              strata = satisfaction, breaks = 4)
estate_train <- training(estate_split)
estate_test <- testing(estate_split)

#TASK 2----

estate_folds <- vfold_cv(estate_train, v = 10, repeats = 5,
                     strata = satisfaction)

save(estate_train, estate_test, estate_folds, file = here("data/estate_split.rda"))
